+++ src/main/java/com/sparta/travelconquestbe/api/bookmark/service/BookmarkService.java
@@ -2,7 +2,6 @@ package com.sparta.travelconquestbe.api.bookmark.service;
 
 import com.sparta.travelconquestbe.api.bookmark.dto.response.BookmarkCreateResponse;
 import com.sparta.travelconquestbe.api.bookmark.dto.response.BookmarkListResponse;
-import com.sparta.travelconquestbe.api.bookmark.dto.response.BookmarkRankingResponse;
 import com.sparta.travelconquestbe.common.auth.AuthUserInfo;
 import com.sparta.travelconquestbe.common.exception.CustomException;
 import com.sparta.travelconquestbe.domain.bookmark.entity.Bookmark;
@@ -11,7 +10,6 @@ import com.sparta.travelconquestbe.domain.route.entity.Route;
 import com.sparta.travelconquestbe.domain.route.repository.RouteRepository;
 import com.sparta.travelconquestbe.domain.user.entity.User;
 import com.sparta.travelconquestbe.domain.user.repository.UserRepository;
-import java.time.YearMonth;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -72,36 +70,5 @@ public class BookmarkService {
     }
 
     bookmarkRepository.delete(bookmark);
-  }
-
-  @Transactional(readOnly = true)
-  public Page<BookmarkRankingResponse> getMonthlyRankings(int year, int month, int page, int size) {
-    validateYearAndMonth(year, month);
-
-    PageRequest pageRequest = PageRequest.of(page - 1, Math.min(size, 10));
-    return bookmarkRepository.findMonthlyRankings(year, month, pageRequest);
-  }
-
-  @Transactional(readOnly = true)
-  public Page<BookmarkRankingResponse> getRealtimeRankings(int page, int size) {
-    PageRequest pageRequest = PageRequest.of(page - 1, Math.min(size, 10));
-    return bookmarkRepository.findRealtimeRankings(pageRequest);
-  }
-
-  @Transactional(readOnly = true)
-  public Page<BookmarkRankingResponse> getAlltimeRankings(int page, int size) {
-    PageRequest pageRequest = PageRequest.of(page - 1, Math.min(size, 10));
-    return bookmarkRepository.findAlltimeRankings(pageRequest);
-  }
-
-  private void validateYearAndMonth(int year, int month) {
-    if (month < 1 || month > 12) {
-      throw new CustomException("BOOKMARK#4_001", "월은 1~12 사이여야 합니다.", HttpStatus.BAD_REQUEST);
-    }
-
-    YearMonth inputDate = YearMonth.of(year, month);
-    if (inputDate.isAfter(YearMonth.now())) {
-      throw new CustomException("BOOKMARK#4_002", "요청하신 날짜는 현재 날짜보다 미래일 수 없습니다.", HttpStatus.BAD_REQUEST);
-    }
   }
 }
